BE/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
llama-index
nest_asyncio
langchain
pyyaml
crewai
requests
beautifulsoup4
duckduckgo-search
crewai-tools
